spark/requirements.txt
pymongo
python-dotenv
google-cloud-bigquery
pandas